Implemented asked for changes by Joyal.
Implemented asked for changes by Joyal.
Made a new file with file name teamData.tsx and kept the data of people in it. Then imported it and used it in the team/page.tsx component by mapping it over.

Also, added a in-place scaling effect in the team's individual's pictures.